Split container module
